tax stuff needs to be updated

+++ src/Traitement_Des_articles/Item.java
@@ -18,7 +18,7 @@ private final int weight;
         return this.getPrice();
     }
     @Override
-    public double taxRatePerTenThousand() {
+    public long taxRatePerTenThousand() {
         return 1000;
     }
 //constructor  with arguments
