+++ src/Traitement_des_factures/Payable.java
@@ -3,6 +3,6 @@ package Traitement_des_factures;
 public interface Payable  {
      String label();
      long cost();
-    double taxRatePerTenThousand();
+    long taxRatePerTenThousand();
 
 }
